requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.6.6
backcall==0.2.0
black==21.7b0
click==8.0.1
coverage==5.5
decorator==5.0.9
Django==3.2.6
django-cors-headers==3.8.0
django-extensions==3.1.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
drf-writable-nested==0.6.3
Faker==8.11.0
ipdb==0.13.9
ipython==7.26.0
ipython-genutils==0.2.0
isort==5.9.3
jedi==0.18.0
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
pylint==2.9.6
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
traitlets==5.0.5
wcwidth==0.2.5
wrapt==1.12.1
